Plot distributions to explore data and/or choose priors.

These are helper functions to help explore data and/or choose priors. Click further for more details.

plotBeta

plotGamma

plotInvGamma

plotLogNormal

plotNormal

plotPareto

plotPoisson

Note

Choosing priors correctly is very important. Please see http://fportman.com/blog/bayesab-0-dot-7-0-plus-a-primer-on-priors/ for a detailed example of choosing priors within bayesAB. Here are some ways to leverage objective/diffuse (assigning equal probability to all values) priors:

Gamma(eps, eps) ~ Gamma(.00005, .00005) will be effectively diffuse

InvGamma(eps, eps) ~ InvGamma(.00005, .00005) will be effectively diffuse

Pareto(eps, eps) ~ Pareto(.005, .005) will be effectively diffuse

Keep in mind that the Prior Plots for bayesTest's run with diffuse priors may not plot correctly as they will not be truncated as they approach infinity. See plot.bayesTest for how to turn off the Prior Plots.

plot... functions are generated programmatically so the function calls in their body will be substituted directly